requirements.txt
requests
networkx
numpy
pandas
scikit-learn
matplotlib
seaborn
scipy
xgboost
tkcalendar
python-dotenv